Upgrade webmock and rollbar-blanket
The cause of churn in this commit is to do with Webmock > 2.0 changing
how it handles Basic auth in the URL:

  > WebMock 2.0 does not match basic authentication credentials in the
  > userinfo part of the url, with credentials passed in Authorization:
  > Basic ... header anymore. It now treats the Authorization header and
  > credentials in the userinfo part of a url as two completely separate
  > attributes of a request. # Please enter the commit message for your
  > changes.

This impacts us b/c we pass the userinfo as part of the URL string in an
ENV Var. But Excon parses that out and sends it as a Basic auth header.
Therefore, the URLs we're stubbing don't match what ends up being used.
Hence, changing how we stub those requests. 😅