add /user/notifications/unread-count endpoint
Most users do not look at their notifications. It is therefore a bit
wasteful of their bandwidth and loading time to load data they're not
necessarily going to look at. The `/user/notifications/unread-count`
does a `COUNT` for the unread notifications for the user, which saves
bandwidth and processing time as they don't have to download the entire
body of the messges. The ruby app should also have to do less work and
not have to encode the body for the JSON messages unless the user is
actively looking at their messages.