log the class name on error
the pliny log scrubber can do fun things with the params passed to
`Pliny.log`. For example, if you pass it a `Sequel` class, the scrubber
could attempt to call various methods. In particular, it might inquire
if the class responds to `to_hash` and if so it might call `to_hash`...
and `User.to_hash` yields `SELECT * FROM "users"` which is bad news for
performance...